refactor(crypto): move crypto config to infrastructure layer
Consolidate crypto configuration files under infrastructure/config directory for better project structure organization
